requirements.txt
altgraph==0.17.3
asgiref==3.5.2
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.2
django-appconf==1.0.5
django-contrib-comments==2.2.0
django-imagekit==4.1.0
django-js-asset==2.0.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.17
future==0.18.2
idna==3.4
importlab==0.8
lxml==4.9.2
networkx==3.0
pefile==2022.5.30
pilkit==2.0
pytz==2023.3
pywin32-ctypes==0.2.0
reportlab==3.6.12
requests==2.28.2
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.3
tzdata==2022.4
urllib3==1.26.14
